// static-project/wsc-pc-vis/client/.yalc/@youzan/ebiz-components/components/easy-form/types.ts
export interface IFieldStatus {
  disabled: boolean;
  visible: boolean;
}
// ...
export type EasyFormWatchFunction<Value = any> = (
  value: Value,
  ctx: IEasyFormWatchCtx,
  easyForm: OriginalFormType
) => void;
// ...
interface IPartialFieldProps {
  className?: string;
  style?: React.CSSProperties;
  withoutLabel?: boolean;
}
// interface IEasyFormHooksLikeWatcher {
//   factory: EasyFormWatchFunction<Record<string, any>>;
//   deps: string[];
// }
export type EasyFormWatchType = /** IEasyFormHooksLikeWatcher |*/ Record<
  string,
  EasyFormWatchFunction<any>
>;
export type EasyFormConfigLabel = ((name: string, index: number) => string) | string;
interface IEasyFormBaseConfig {
  name: string;
  label?: EasyFormConfigLabel;
  type: FormFieldsType;
  visible?: boolean;
  onChange?(input: any): void;
  /**
   * field组件属性
   */
  disabled?: boolean;
  required?: boolean;
  defaultValue?: any;
  validators?: FieldModel<any>['validators'];
  /** 配置校验时机 */
  validateOccasion?: ValidateOccasion;
  /**
   * field属性，这个属性将不会覆盖配置中声明的属性，如required
   */
  fieldProps?: IPartialFieldProps & Record<string, any>;
  /** 传递给组件的属性 */
  inheritProps?: Record<string, any>;
  /**
   * 在field前面插入内容，如：
   * `label: {{prefix}} [placeholder   ] {{suffix}}`
   *
   * 可以用构造形如`[placeholder  ] 至 [placeholder  ]`这样的范围输入框
   */
  prefix?: React.ReactNode;
  suffix?: React.ReactNode;

  /**
   * 在field底部显示的提示文字，如
   * ```
   * label: [placeholder   ]
   *        查看ebiz-component文档，请访问<a>这里</a>
   * ```
   */
  helpDesc?: React.ReactNode;

  /**
   * 监听字段的变化去改变自身，有两种用法：
   * 1. [**TODO**]~~一种是类似hooks的调用方式：为一个接收两个参数的函数~~
   * 2. 另一种是直接对某个field的值进行订阅：定义为一个对象，对象的键名则是监听的目标
   */
  watch?: EasyFormWatchType;
}

interface INoValueProps<Interface, ValueType = any> {
  defaultValue?: ValueType;
  inheritProps?: Omit<Interface, 'value' | 'defaultValue' | 'disabled' | 'onChange' | 'onBlur'>;
}
interface IEasyFormInputConfig extends INoValueProps<Partial<IInputCoreProps>, string> {
  type: 'Input';
}
interface IEasyFormNumberInputConfig extends INoValueProps<Partial<INumberInputProps>> {
  type: 'NumberInput';
}
interface IOption {
  text: string;
  value: string;
  [key: string]: any;
}
interface IEasyFormSelectConfig extends INoValueProps<Omit<ISelectProps, 'data'>, string> {
  type: 'Select';
  options: IOption[];
}
interface IEasyFormCheckboxConfig extends INoValueProps<ICheckboxGroupProps<string>, string[]> {
  type: 'Checkbox';
  options: IOption[];
}
interface IEasyFormRadioConfig extends INoValueProps<IRadioGroupProps<any>> {
  type: 'Radio';
  options: IOption[];
}
/**
 * 需要options属性的类型
 */
export type EasyFormOptionsRequiredTypes = (
  | IEasyFormSelectConfig
  | IEasyFormCheckboxConfig
  | IEasyFormRadioConfig
) &
  IEasyFormBaseConfig;
interface IEasyFormSwitchConfig extends INoValueProps<Omit<ISwitchProps, 'checked'>, string> {
  type: 'Switch';
}

/**
 * ColorPicker与zent不同，通过displayType来分发不同的ColorPicker组件
 */
export interface IEasyFormExtendsColorPickerProps {
  displayType?: 'board' | 'simple';
}
interface IEasyFormColorPickerConfig
  extends INoValueProps<IColorPickerProps & IEasyFormExtendsColorPickerProps> {
  type: 'ColorPicker';
}
interface IEasyFormDatePickerConfig extends INoValueProps<IDatePickerProps, [string, string]> {
  type: DatePickerType;
}
interface IModifiedPresetDays {
  text: string;
  value: any;
}
interface IEasyFormDateRangePickerConfig
  extends INoValueProps<
    Omit<IDateRangeQuickPickerProps, 'onChange' | 'format' | 'preset'>,
    [[string, string], any]
  > {
  type: 'DateRangeQuickPicker';
  preset?: IModifiedPresetDays[];
}
interface IEasyFormCustomConfig {
  type: 'Custom';
  renderField: React.ReactNode;
}
interface IForbidden_InternalGroupConfig extends IGroupEasyFormConfigs {
  /**
   * 请不要直接指定这个类型，如果要将config进行分组，
   * 请使用`EasyForm.group`方法将需要的config进行
   * 分组
   * ```
   * const { group } = EasyForm
   * return group({
   *  groupName: 'group1',
   *  config: [
   *    ...fieldConfig
   *  ]
   * })
   * ```
   */
  type: '__internal_group__';
  label?: string;
}
interface IForbidden_InternalListConfig extends IListEasyFormConfigs {
  /**
   * 请不要直接指定这个类型，如果要将config进行分组，
   * 请使用`EasyForm.list`方法将需要的config进行
   * 分组
   * ```
   * const { list } = EasyForm
   * return list({
   *  groupName: 'group1',
   *  repeatConfig: fieldConfig
   * })
   * ```
   */
  type: '__internal_list__';
}
type EasyFormPlainType = {
  type: 'Plain';
  node: React.ReactNode;
} & IEasyFormBaseConfig;
export type EasyFormInternalListType = IForbidden_InternalListConfig & IEasyFormBaseConfig;
export type EasyFormInternalGroupType = IForbidden_InternalGroupConfig & IEasyFormBaseConfig;
export type EasyFormConfigValidType = (
  | IEasyFormInputConfig
  | IEasyFormNumberInputConfig
  | IEasyFormSelectConfig
  | IEasyFormCheckboxConfig
  | IEasyFormRadioConfig
  | IEasyFormCustomConfig
  | IEasyFormSwitchConfig
  | IEasyFormColorPickerConfig
  | IEasyFormDatePickerConfig
  | IEasyFormDateRangePickerConfig
) &
  IEasyFormBaseConfig;
export type EasyFormConfigType =
  | EasyFormPlainType
  | EasyFormConfigValidType
  | EasyFormInternalGroupType
  | EasyFormInternalListType;

export interface IGroupEasyFormConfigs {
  groupName: string;
  groupTitle: string | React.ReactNode;
  /**
   * 是否折叠分组
   */
  collapse?: boolean;
  config: EasyFormConfigType[];
  onChange?(value: Record<string, any>): void;
  /** 是否整块区域不可用，如果这个值被修改为false，则disabled属性会传到各个field上 */
  disabled?: boolean;
  watch?: EasyFormWatchType;
  hideGroupHeader?: boolean;
}

export interface IEasyFormListRepeatMethods {
  /** 添加一项，并将repeatConfig.defaultValue设置为初始值 */
  add(): void;
  /** 添加n个项，如果填写了`defaultValue`，就会将这个入参作为初始值来初始化 */
  add(addCount?: number, defaultValue?: any): void;
  /** 删除一项 */
  delete(): void;
  /** 删除n项，如果填写了`startIndex`，就会从`startIndex`开始删除n项 */
  delete(deleteCount?: number, startIndex?: number): void;
}

export interface IEasyFormRepeatTriggerProps {
  methods: IEasyFormListRepeatMethods;
  required: boolean;
  visible: boolean;
  disabled: boolean;
  label: string | null;
}
export interface IListEasyFormConfigs {
  listName: string;
  label?: string;
  fieldProps?: IPartialFieldProps & Record<string, any>;
  /**
   * 渲染触发数组变化的节点，该方法会带有几个实用方法，用于
   * 控制`repeatConfig`渲染的数量：
   * 1. `add`     添加一个或n个`repeatConfig`
   * 2. `delete`  删除最后一个或者删除第n个`repeatConfig`
   * 3. `filter`  过滤`repeatConfig`
   *
   * **请注意，一定要在返回的组件中，预留children的位置以供
   * 渲染list的内容**
   */
  repeatTrigger: React.ComponentType<IEasyFormRepeatTriggerProps>;
  /**
   * 需要重复的field配置，**注意：在`list`中，因为所有数据列
   * 表都是通过这个属性来渲染的，所以不允许使用`watch`方法来监
   * 听其他数据。**
   */
  repeatConfig: EasyFormConfigType;
  /**
   * 在重复生成`repeatConfig`的时候，需要指定唯一id
   * 如果没有声明这个方法，会采用`name + index`的方式
   * 命名每一个key
   */
  getRepeatKey?(name: string, idx: number): any;
  onChange?(value: any[]): void;
  /** 如果这个值为false，则表示包括，trigger在内都不会显示 */
  visible?: boolean;
  /** 这个参数将会影响repeatTrigger的disabled参数 */
  disabled?: boolean;
  watch?: EasyFormWatchType;
}
// ...
export type NormalEasyFormModel = FieldModel<any>;
export type GroupEasyFormModel = FieldSetModel<Record<string, NormalEasyFormModel>>;
export type ListEasyFormModel = FieldArrayModel<any>;
export type EasyFormModelType = NormalEasyFormModel | GroupEasyFormModel | ListEasyFormModel;
// status type
export type NormalStatusModel = FieldModel<IFieldStatus>;
export type GroupStatusModel = FieldSetModel<Record<string, NormalStatusModel>>;
export type ListStatusModel = NormalStatusModel;
export type EasyFormStatusModelType = NormalStatusModel | GroupStatusModel | ListStatusModel;